silx.gui.plot.backends.glutils.GLPlotFrame: move PlotAxis in its own file
